+++ lib/glimmer/dsl/swt/dsl.rb
@@ -26,7 +26,6 @@ module Glimmer
           data_binding
           cursor
           font
-          image
           property
           block_property
           widget