+++ lib/glimmer/dsl/swt/img_expression.rb
@@ -1,4 +1,5 @@
-require 'glimmer/dsl/expression'
+require 'glimmer/dsl/static_expression'
+require 'glimmer/dsl/top_level_expression'
 require 'glimmer/swt/image_proxy'
 
 module Glimmer
@@ -6,11 +7,8 @@ module Glimmer
     module SWT
       # image expression
       # Note: Cannot be a static expression because it clashes with image property expression
-      class ImageExpression < Expression
-        def can_interpret?(parent, keyword, *args, &block)
-          keyword.to_s == 'image' and
-            (parent.nil? or !parent.respond_to?('image'))
-        end
+      class ImgExpression < StaticExpression
+        include TopLevelExpression
   
         def interpret(parent, keyword, *args, &block)
           Glimmer::SWT::ImageProxy.new(*args)