+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 - Set marginLeft marginRight marginTop marginBottom to 0 when methods are available on a layout (like RowLayout)
 - Add :no_margin SWT style to layouts
 - Add :no_margin SWT style to composites (affecting their default layout)
+- Accept `ImageProxy` as arg for image method
+- `img` keyword for instantiating an SWT Image object
 
 **CI:**